refactor(stringCalculator): Custom, Default의 구분을 instanceof 로 변경
모든 CustomDelimiter들이 isDefault를 구현한다는 의미가 이상하여 변경
