+++ src/main/java/stringcalculator/domain/Delimiter/DefaultDelimiter.java
@@ -1,12 +1,7 @@
-package stringcalculator.domain;
+package stringcalculator.domain.Delimiter;
 
 public class DefaultDelimiter implements Delimiter {
     private final String DEFAULT_DELIMITER = ",|:";
-
-    @Override
-    public boolean isDefault() {
-        return true;
-    }
 
     @Override
     public String[] split(String input) {
