+++ src/main/java/stringcalculator/domain/Delimiter/DelimiterFactory.java
@@ -1,4 +1,4 @@
-package stringcalculator.domain;
+package stringcalculator.domain.Delimiter;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
